Source formatting (remove trailing blank line on some source)
LEP-1963

git-svn-id: svn://svn.liferay.com/repos/public/portal/trunk@3580 05bdf26c-840f-0410-9ced-eb539d925f36

+++ documentlibrary-ejb/src/com/liferay/documentlibrary/util/IndexerImpl.java
@@ -139,9 +139,7 @@ public class IndexerImpl {
 				LuceneFields.getKeyword(LuceneFields.COMPANY_ID, companyId));
 			doc.add(
 				LuceneFields.getKeyword(LuceneFields.PORTLET_ID, portletId));
-			doc.add(
-				LuceneFields.getKeyword(LuceneFields.GROUP_ID,
-					groupId.longValue()));
+			doc.add(LuceneFields.getKeyword(LuceneFields.GROUP_ID, groupId));
 
 			doc.add(LuceneFields.getFile(LuceneFields.CONTENT, is, fileExt));
